+++ IntermediateCodeGenerator/intermediateGen.cpp
@@ -1,0 +1,1 @@
+#include "intermediateGen.hpp"
