--- Lexical Analysis/test.cor
@@ -1,1 +1,0 @@
-var:int a = 0;
