+++ main.cpp
@@ -1,5 +1,5 @@
-#include "./Preprocessor/preprocessor.cpp"
-#include "./Syntax Analysis/syntaxAnalysis.cpp"
+#include "./Preprocessor/preprocessor.hpp"
+#include "./SyntaxAnalysis/syntaxAnalysis.hpp"
 
 #include <iostream>
 #include <string>
@@ -8,17 +8,17 @@ using namespace std;
 int main(int argc, char* argv[]) {
     if (argc != 2) {
         cout << "Format: erosion <filename>\n";
-        return 0;
+        return 1;
     }
 
     string extension = string(argv[1]);
     if (extension.size() < 4) {
         cout << "Error: <filename> must be a .cor file\n";
-        return 0;
+        return 1;
     }
     if (extension.substr(extension.size() - 4, 4) != ".cor") {
         cout << "Error: <filename> must be a .cor file\n";
-        return 0;
+        return 1;
     }
 
     Preprocessor* preprocessor = new Preprocessor(argv[1]);
